[svn] Allow Wget to be compiled with a K&R compiler. Published in <sxszo0af9k6.fsf@florida.arsdigita.de>.

+++ src/cookies.c
@@ -84,7 +84,7 @@ cookie_new (void)
 
   /* If we don't know better, assume cookie is non-permanent and valid
      for the entire session. */
-  cookie->expiry_time = ~0UL;
+  cookie->expiry_time = ~(unsigned long)0;
 
   /* Assume default port. */
   cookie->port = 80;
